Refactor AccountBalances comp to simplify code
