+++ client/components/account-balances/test/index.test.tsx
@@ -7,24 +7,10 @@ import { render, screen, fireEvent, within } from '@testing-library/react';
 /**
  * Internal dependencies
  */
-import AccountBalances from '../';
-import AccountBalancesTabPanel from '../balances-tab-panel';
+import AccountBalances from '..';
 import { getCurrencyTabTitle } from '../utils';
-import { getGreeting } from 'components/account-overview-header/utils';
-import { useCurrentWpUser } from 'components/account-overview-header/hooks';
 import { useAllDepositsOverviews, useInstantDeposit } from 'wcpay/data';
 import { useSelectedCurrency } from 'wcpay/overview/hooks';
-
-const mockUser = {
-	id: 123,
-	first_name: 'Tester',
-	username: 'admin',
-	name: 'admin',
-	nickname: 'Tester-nickname',
-	last_name: 'Tester-lastname',
-	email: 'tester@test.com',
-	locale: 'en',
-};
 
 const mockAccount: AccountOverview.Account = {
 	default_currency: 'USD',
@@ -72,14 +58,6 @@ jest.mock( '../utils', () => ( {
 	getCurrencyTabTitle: jest.fn(),
 } ) );
 
-jest.mock( 'components/account-overview-header/utils', () => ( {
-	getGreeting: jest.fn(),
-} ) );
-
-jest.mock( 'components/account-overview-header/hooks', () => ( {
-	useCurrentWpUser: jest.fn(),
-} ) );
-
 jest.mock( 'wcpay/data', () => ( {
 	useAllDepositsOverviews: jest.fn(),
 	useInstantDeposit: jest.fn(),
@@ -89,12 +67,6 @@ jest.mock( 'wcpay/overview/hooks', () => ( {
 	useSelectedCurrency: jest.fn(),
 } ) );
 
-const mockGetGreeting = getGreeting as jest.MockedFunction<
-	typeof getGreeting
->;
-const mockUseCurrentWpUser = useCurrentWpUser as jest.MockedFunction<
-	typeof useCurrentWpUser
->;
 const mockGetCurrencyTabTitle = getCurrencyTabTitle as jest.MockedFunction<
 	typeof getCurrencyTabTitle
 >;
@@ -194,34 +166,12 @@ describe( 'AccountBalances', () => {
 		global.wcpaySettings = mockWcPaySettings;
 	} );
 
-	test( 'renders', () => {
-		const expectedGreeting = 'Good afternoon, Tester 👋';
-		mockGetGreeting.mockReturnValue( expectedGreeting );
-		mockUseCurrentWpUser.mockReturnValue( {
-			user: mockUser,
-			isLoading: false,
-		} );
-		mockGetCurrencyTabTitle.mockReturnValue( 'USD Balance' );
-		mockOverviews( [ createMockOverview( 'usd', 100, 200, 0 ) ] );
-
-		const { container } = render( <AccountBalances /> );
-		expect( container ).toMatchSnapshot();
-	} );
-} );
-
-describe( 'AccountBalancesTabPanel', () => {
-	beforeEach( () => {
-		global.wcpaySettings = mockWcPaySettings;
-	} );
-
 	test( 'renders the correct tab title and currency data', () => {
 		mockGetCurrencyTabTitle.mockReturnValue( 'USD Balance' );
 		mockOverviews( [ createMockOverview( 'usd', 10000, 20000, 0 ) ] );
 
 		// Use a query method returned by the render function: (you could also use `container` which will represent `document`)
-		const { getByText, getByLabelText } = render(
-			<AccountBalancesTabPanel />
-		);
+		const { getByText, getByLabelText } = render( <AccountBalances /> );
 
 		// Check the tab title is rendered correctly.
 		getByText( 'Available funds' );
@@ -239,9 +189,7 @@ describe( 'AccountBalancesTabPanel', () => {
 		mockGetCurrencyTabTitle.mockReturnValue( 'JPY Balance' );
 		mockOverviews( [ createMockOverview( 'jpy', 12300, 4560, 0 ) ] );
 
-		const { getByText, getByLabelText } = render(
-			<AccountBalancesTabPanel />
-		);
+		const { getByText, getByLabelText } = render( <AccountBalances /> );
 
 		// Check the tab title is rendered correctly.
 		getByText( 'Available funds' );
@@ -271,9 +219,7 @@ describe( 'AccountBalancesTabPanel', () => {
 			setSelectedCurrency: mockSetSelectedCurrency,
 		} );
 
-		const { getByLabelText, getByRole } = render(
-			<AccountBalancesTabPanel />
-		);
+		const { getByLabelText, getByRole } = render( <AccountBalances /> );
 
 		// Check the active tab is rendered correctly.
 		getByRole( 'tab', {
@@ -306,9 +252,7 @@ describe( 'AccountBalancesTabPanel', () => {
 			setSelectedCurrency: mockSetSelectedCurrency,
 		} );
 
-		const { getByLabelText, getByRole } = render(
-			<AccountBalancesTabPanel />
-		);
+		const { getByLabelText, getByRole } = render( <AccountBalances /> );
 
 		// Check the default active tab is rendered correctly.
 		getByRole( 'tab', {
@@ -336,7 +280,7 @@ describe( 'AccountBalancesTabPanel', () => {
 			createMockOverview( 'jpy', 2000, 9000, 0 ),
 		] );
 
-		const { getByLabelText } = render( <AccountBalancesTabPanel /> );
+		const { getByLabelText } = render( <AccountBalances /> );
 
 		// Get all the tab elements to check the tab titles are rendered correctly and for testing tab switching.
 		const tabTitles = screen.getAllByRole( 'tab' );
@@ -382,7 +326,7 @@ describe( 'AccountBalancesTabPanel', () => {
 
 	test( 'renders the correct tooltip text for the available balance', () => {
 		mockOverviews( [ createMockOverview( 'usd', 10000, 20000, 0 ) ] );
-		render( <AccountBalancesTabPanel /> );
+		render( <AccountBalances /> );
 
 		// Check the tooltips are rendered correctly.
 		const tooltipButton = screen.getByRole( 'button', {
@@ -400,7 +344,7 @@ describe( 'AccountBalancesTabPanel', () => {
 
 	test( 'renders the correct tooltip text for a negative available balance', () => {
 		mockOverviews( [ createMockOverview( 'usd', 10000, -20000, 0 ) ] );
-		render( <AccountBalancesTabPanel /> );
+		render( <AccountBalances /> );
 
 		// Check the tooltips are rendered correctly.
 		const tooltipButton = screen.getByRole( 'button', {
@@ -419,7 +363,7 @@ describe( 'AccountBalancesTabPanel', () => {
 
 	test( 'renders the correct tooltip text for a negative pending balance', () => {
 		mockOverviews( [ createMockOverview( 'usd', -10000, 20000, 0 ) ] );
-		render( <AccountBalancesTabPanel /> );
+		render( <AccountBalances /> );
 
 		// Check the tooltips are rendered correctly.
 		const tooltipButton = screen.getByRole( 'button', {
@@ -439,7 +383,7 @@ describe( 'AccountBalancesTabPanel', () => {
 	test( 'renders the correct tooltip text for the pending balance', () => {
 		const delayDays = mockAccount.deposits_schedule.delay_days;
 		mockOverviews( [ createMockOverview( 'usd', 10000, 20000, 0 ) ] );
-		render( <AccountBalancesTabPanel /> );
+		render( <AccountBalances /> );
 
 		// Check the tooltips are rendered correctly.
 		const tooltipButton = screen.getByRole( 'button', {
@@ -460,7 +404,7 @@ describe( 'AccountBalancesTabPanel', () => {
 
 	test( 'renders instant deposit button correctly', () => {
 		mockOverviews( [ createMockOverview( 'usd', 10000, 20000, 30000 ) ] );
-		render( <AccountBalancesTabPanel /> );
+		render( <AccountBalances /> );
 
 		screen.getByRole( 'button', {
 			name: 'Deposit available funds',
@@ -469,7 +413,7 @@ describe( 'AccountBalancesTabPanel', () => {
 
 	test( 'does not render instant deposit button when instant amount is 0', () => {
 		mockOverviews( [ createMockOverview( 'usd', 10000, 20000, 0 ) ] );
-		render( <AccountBalancesTabPanel /> );
+		render( <AccountBalances /> );
 
 		expect(
 			screen.queryByRole( 'button', { name: 'Deposit available funds' } )
@@ -480,7 +424,7 @@ describe( 'AccountBalancesTabPanel', () => {
 		const mockOverview = createMockOverview( 'usd', 10000, 20000, 0 );
 		mockOverview.instant = undefined;
 		mockOverviews( [ mockOverview ] );
-		render( <AccountBalancesTabPanel /> );
+		render( <AccountBalances /> );
 
 		expect(
 			screen.queryByRole( 'button', { name: 'Deposit available funds' } )
